Refatoração dos códigos de calculo do pedido + oraganização em pastas
