+++ src/main/java/br/com/alura/comex/pedido/CalculosPedidos.java
@@ -1,35 +1,17 @@
-package br.com.alura.comex;
+package br.com.alura.comex.pedido;
 
 import java.math.BigDecimal;
-import java.time.LocalDate;
 
-public class Pedido {
+import br.com.alura.comex.formatador.Formatacoes;
 
-	private String categoria;
-	private String produto;
-	private String cliente;
-
-	private BigDecimal preco;
-	private int quantidade;
-
-	private BigDecimal montanteDeVendas = BigDecimal.ZERO;
-	private Pedido pedidoMaisBarato = null;
-	private Pedido pedidoMaisCaro = null;
+public class CalculosPedidos {
+	
+	protected BigDecimal montanteDeVendas = BigDecimal.ZERO;
+	protected Pedido pedidoMaisBarato = null;
+	protected Pedido pedidoMaisCaro = null;
+	
 	private Formatacoes formatador = new Formatacoes();
-	private LocalDate data;
-
-	public Pedido() {
-	}
-
-	public Pedido(String categoria, String produto, String cliente, BigDecimal preco, int quantidade, LocalDate data) {
-		this.categoria = categoria;
-		this.produto = produto;
-		this.cliente = cliente;
-		this.preco = preco;
-		this.quantidade = quantidade;
-		this.data = data;
-	}
-
+	
 	public boolean isMaisBaratoQue(Pedido pedidoAtual) {
 		if (this.pedidoMaisBarato == null || pedidoAtual.getPreco()
 				.multiply(new BigDecimal(pedidoAtual.getQuantidade())).compareTo(this.pedidoMaisBarato.getPreco()
@@ -55,11 +37,11 @@ public class Pedido {
 	public String pedidoMaisBaratoFormatado() {
 		return formatador.formatarPedidoMaisBarato(this.pedidoMaisBarato);
 	}
-
+	
 	public Pedido getPedidoMaisBarato() {
 		return this.pedidoMaisBarato;
 	}
-
+	
 	public String pedidoMaisCaroQueFormatado() {
 		return formatador.formatarPedidoMaisCaro(this.pedidoMaisCaro);
 	}
@@ -77,35 +59,4 @@ public class Pedido {
 		return this.montanteDeVendas = this.montanteDeVendas
 				.add(pedidoAtual.getPreco().multiply(new BigDecimal(pedidoAtual.getQuantidade())));
 	}
-
-	public String getCategoria() {
-		return categoria;
-	}
-
-	public String getProduto() {
-		return produto;
-	}
-
-	public String getCliente() {
-		return cliente;
-	}
-
-	public BigDecimal getPreco() {
-		return preco;
-	}
-
-	public int getQuantidade() {
-		return quantidade;
-	}
-
-	public LocalDate getData() {
-		return data;
-	}
-
-	@Override
-	public String toString() {
-		return "Pedido{" + "categoria='" + categoria + '\'' + ", produto='" + produto + '\'' + ", cliente='" + cliente
-				+ '\'' + ", preco=" + preco + ", quantidade=" + quantidade + ", data=" + data + '}';
-	}
-
 }
